chore: remove JumpStartPredictor
