+++ tests/unit/sagemaker/jumpstart/test_predictor.py
@@ -3,7 +3,8 @@ from __future__ import absolute_import
 from unittest.mock import patch
 from sagemaker.deserializers import JSONDeserializer
 from sagemaker.jumpstart.enums import MIMEType
-from sagemaker.jumpstart.predictor import JumpStartPredictor
+
+from sagemaker import predictor
 
 
 from sagemaker.jumpstart.utils import verify_model_region_and_return_specs
@@ -23,12 +24,14 @@ def test_list_jumpstart_scripts(
     patched_get_model_specs.side_effect = get_special_model_spec
 
     # version not needed for JumpStart predictor
-    model_id = "predictor-specs-model"
+    model_id, model_version = "predictor-specs-model", "*"
 
-    predictor = JumpStartPredictor(endpoint_name="blah", model_id=model_id)
+    js_predictor = predictor.retrieve_default(
+        endpoint_name="blah", model_id=model_id, model_version=model_version
+    )
 
-    assert predictor.content_type == MIMEType.X_TEXT
-    assert isinstance(predictor.serializer, IdentitySerializer)
+    assert js_predictor.content_type == MIMEType.X_TEXT
+    assert isinstance(js_predictor.serializer, IdentitySerializer)
 
-    assert isinstance(predictor.deserializer, JSONDeserializer)
-    assert predictor.accept == MIMEType.JSON
+    assert isinstance(js_predictor.deserializer, JSONDeserializer)
+    assert js_predictor.accept == MIMEType.JSON
